fix appned in csv

+++ raspi/DataLogger.py
@@ -1,26 +1,34 @@
 #!/usr/bin/python3
 
+import os.path
 import csv
 from Logger import log
 
 fileloc = "./data.csv"
 
 
-def writeData():
+def writeDataSamples():
     writer.writeheader()
-    writer.writerow(
-        {'timestamp': 0,
-         'temperature': 10,
-         'humidity': 2})
-    writer.writerow({'timestamp': 1, 'temperature': 12, 'humidity': 3})
+    writer.writerow({'timestamp': 0,
+                     'temperature': 0,
+                     'humidity': 0,
+                     "conductivity": 0.00})
 
 
 def writeDict(dict):
     log("Writing values to csv...")
-    writer.writerow(dict)
+    with open('document.csv', 'a') as data_csv:
+        fieldnames = ['timestamp', 'temperature', 'humidity', "conductivity"]
+        writer = csv.DictWriter(data_csv, fieldnames=fieldnames)
+        writer.writerow(dict)
 
 
-with open(fileloc, mode='w') as data_csv:
-    fieldnames = ['timestamp', 'temperature', 'humidity']
-    writer = csv.DictWriter(data_csv, fieldnames=fieldnames)
-    writeData()
+if os.path.isfile(fileloc):
+    print ("csv data file exist")
+else:
+    print ("csv data file not exist, create...")
+    with open(fileloc, mode='w') as data_csv:
+        fieldnames = ['timestamp', 'temperature', 'humidity', "conductivity"]
+        writer = csv.DictWriter(data_csv, fieldnames=fieldnames)
+        writeDataSamples()
+
